moved logger to separate module
