+++ logger/build.gradle.kts
@@ -1,0 +1,7 @@
+dependencies {
+    implementation("org.aspectj:aspectjweaver:1.9.20.1")
+}
+
+tasks.bootJar {
+    enabled = false
+}
