+++ logger/src/main/java/executor/service/logger/aspect/LoggingAspect.java
@@ -1,4 +1,4 @@
-package executor.service.publisher.logger;
+package executor.service.logger.aspect;
 
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.JoinPoint;
