+++ logger/src/main/java/executor/service/logger/config/LoggerConfiguration.java
@@ -1,0 +1,12 @@
+package executor.service.logger.config;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Bean;
+
+public class LoggerConfiguration {
+    @Bean
+    public Logger logger() {
+        return LoggerFactory.getLogger("QUEUE_LOGGER");
+    }
+}
