+++ settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "publisher_microservice"
+include("logger")
