+++ src/main/java/executor/service/publisher/config/BeanConfiguration.java
@@ -6,8 +6,6 @@ import executor.service.publisher.queue.proxy.ProxySourceQueueHandlerImpl;
 import executor.service.publisher.queue.scenario.ScenarioSourceQueueHandler;
 import executor.service.publisher.queue.scenario.ScenarioSourceQueueHandlerImpl;
 import okhttp3.OkHttpClient;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -29,8 +27,5 @@ public class BeanConfiguration {
 
     }
 
-    @Bean
-    public Logger logger() {
-        return LoggerFactory.getLogger("QUEUE_LOGGER");
-    }
+
 }
